fix group view download
